Add support for shortener + fixed typos
Add the service to 'shorteners' and add the code after the upload service result parsing.

Should work identically to upload services, for the most part. waaai has an option to display some common file extensions at the end of the url, so I'm using it. If you don't want it, just remove all of the lines with 'extension'.

Sorry for the shitty coding style, tell me if there's anything I need to fix.